Grab ball will take a photo
Finally fixed Git ,this project only can use Github to push ,  git is  unable to access this project, very strange .

This branch was used to test the photo-taking functionality, but it is no longer needed for planned changes outside of the photo-taking functionality, so it has been deprecated. This branch contains some test scenes and cameras (balls).

About the photo function:

Picking up the ball in the scene will automatically take a picture, and the picture will be saved in the running memory, which can be found through the log information address. The output system has not been produced yet.