Dialogue System Demo version
This system uses a database and serialization to create a simple version of a dialogue system. Different dialogue databases can be established to meet various scenario needs. The detailed operations are written in Assets/Dialogue_System/Readme.txt.

This system is very simple and can be used to create different dialogue segments. The drawback is its poor maintainability, and it can be difficult to input large amounts of text.

To see the running demo, simply open the Dialogue_System scene.